add code

+++ week4/up_down.py
@@ -1,9 +1,9 @@
 """up_down"""
 def main():
     """up_down"""
-    num1 = int(input())
-    num2 = int(input())
-    num3 = int(input())
+    num1 = float(input())
+    num2 = float(input())
+    num3 = float(input())
 
     if num1 < num2 < num3:
         print("increasing")
